Checks for existing BYOC AdminAccess role, and deletes it before creating
Checks for the existence of the BYOC AdminAccess role before attempting to
create it.  If it exists (ie: the returned struct is not empty), then lists
the attached role policies, and detaches them, and then deletes the role.
(The role cannot be deleted with attached policies.)

After deleting the role, the contoller then continues to create the role fresh.

Signed-off-by: Christopher Collins <collins.christopher@gmail.com>