Handles unexpected errors returned by the AWS API
Currently, if there is an error creating the role, and the error is an AWS API error (awserr.Error), only "EntityAlreadyExists" is handled, and all other errors are slilently dropped on the floor.  This adds an `else` case to return the awserr.Error.Code() string as an error if it isn't the EntityAlreadyExists error.

REF: panics[1] due to role being nil, ultimately tracked down to unhandled "MalformedPolicyDocument" errors returned by AWS that were being ignored by the code.

[1] https://github.com/openshift/aws-account-operator/pull/300#pullrequestreview-369920242

Signed-off-by: Christopher Collins <collins.christopher@gmail.com>