Splits test-ccs into create-ccs and delete-ccs targets
This breaks the long string of targets in `test-ccs` into `create-ccs` and `delete-ccs` targets.  This will allow us to spin up a CCS cluster by calling the `create-ccs` target directly, for testing things manually when necessary.  `delete-ccs` will handle cleanup.  This also matches the behavior for `test-account`.

The `test-ccs` target still behaves as it did before, with `create-ccs` and `delete-ccs` referencing the same targets that were originally there.

Signed-off-by: Christopher Collins <collins.christopher@gmail.com>