DRY MetricsCollector Describe and Collect
When updating the list of Collectors in MetricsCollector, it is
necessary to make sure that MetricsCollector's Describe and Collect
methods invoke the respective method for each wrapped Collector. This is
slightly tedious and error-prone.

This commit uses some reflection to abstract that part away, so that
adding a new field to MetricsCollector only requires you to do the
things specific to that field.