Metrics for API Calls (1 of 2: kubeclient)
A) Register a new metric, `aws_account_operator_api_request_duration_seconds`,
which times API calls made by the operator. Since we're using Histogram,
we get a `_count` as well. The metric has the following dimensions:
- `controller`: The name of the controller making the API call.
- `method`: The HTTP method (`GET`, `PUT`, etc.).
- `resource`: A normalized interpretation of the target of the API call.
For example, unique resource names are scrubbed. This is so we don't end
up with an explosion of (not very useful) dimension values.
- `status`: The HTTP response status (`200 OK`, `404 Not Found`, etc.).

B) Instrument the controller-runtime Client used by each controller to
publish this metric. This gives us durations and counts for calls to
kubeclient APIs.

Note: we're only doing this for true controllers, not for secretwatcher
or totalaccountwatcher.

Note: This commit only handles kubeclient API calls. A subsequent commit
will instrument the AWS client in a similar fashion.

Jira: [OSD-4390](https://issues.redhat.com/browse/OSD-4390)