Metrics for API Calls (2 of 2: AWS)
Instrument AWS API calls to report duration and count metrics in a
fashion similar to what #410 did for kubeclient calls.

At this time the `resource` label gets assigned the host, which
indicates which AWS service we hit. In the future we may want to drill
into the payloads to report the operation more specifically.

Jira: [OSD-4507](https://issues.redhat.com/browse/OSD-4507)