API call metrics: Defend against `nil` Response
In the kube API call path, we're only reporting metrics if the request
"succeeds", meaning that it returns a real Response (even if that
Response is 4xx/5xx). But in the AWS call path, we may still encounter a
request failure such that the Response is `nil`. This was panicking the
metrics reporting where we fill in the `status` label with the
response's HTTP status.

With this commit, we defend against that by checking for a `nil`
response and registering `{ERROR}` in the `status` label.