Fix AWS API call metrics
Previously we were doing two rather shaky things to time AWS API calls:
- Using a map to keep track of Request start and end times. This is
because one client gets reused for multiple calls. Problem: maps aren't
threadsafe, and we weren't locking.
- Using the address of the Request object as the map key. This is pretty
icky for several reasons, not least of which is that it relies on the
*same* Request object being used for the various parts of the process.
That's likely, but certainly not guaranteed.

Live testing showed some anomalies where we were recording call
durations in the billions of seconds, which was obviously bogus, and
likely caused by one or both of the above.

With this commit, we take advantage of the `Request.Time` field set by
the AWS SDK when the Request object is initialized, and just add a
handler for the last phase of the request that uses the time since that
point for the metric. No more maps, no more weird keys, no more
threading issues (assuming the AWS client itself is threadsafe).

Jira: [OSD-4763](https://issues.redhat.com/browse/OSD-4763)