Consolidates Readme and creates Docs dir
Consolidates the Readme to be a bit more concise and have a quickstart
guide, while moving the more in-depth documentation into a docs folder.

I tried not to modify much of the existing documentation except for the
first `Installation` page mainly to clean up where I had to cut out
stuff for the quickstart.