Refactor mocks to avoid circular imports
We hadn't regenerated mocks since the last time the awsclient was
modified. Doing so as is resulted in a circular import (code in
`awsclient` was referring to `mock` which refers to `awsclient`).

This commit moves the mock-related `IBuilder` implementation into the
`mock` package to obviate this circular import.

It also syncs the `go:generate` directive and the generated file's name
as `zz_generated.mock_client.go`. (The `zz_generated` prefix, in
addition to being a nice eyecatcher, will cause the generated file to be
ignored for code coverage reporting when that's put in place later.)