Remove 'auto author assign' workflow
This workflow configuration, introduced via #526, was empty, meaning it
would always fail
([example](https://github.com/2uasimojo/aws-account-operator/actions/runs/883561188).

We could populate it based on the
[docs](https://github.com/marketplace/actions/auto-author-assign); but I
contend (and, by this commit, assert) that it's not a useful action. If
I've switched off the
[setting](https://github.com/settings/notifications) that notifies me by
default on activity on a PR I've authored, it's because I don't want to
be notified. If I want to make an exception, I'll assign myself.
Otherwise, it ain't right for the repository to make that switch for me.