Add depandabot configuration
Whenever there is a new version of the 3rd party dependencies
used by the project (namely whatever is inside go.mod and go.sum)
then let the project update itself without developer's intervention.

Do the same for the APIs, as now, it's a separate isolated pkg.

The configuration makes sure this check runs on weekly basis.

Note: To make this work you need an admin for AAO to install
      https://github.com/marketplace/dependabot-preview

REF: OSD-6207