Adds post-deployment validation
This script adds the ability to quickly validate an AWS Account Operator
deployment to various environments using OCM, to validate the same steps
a customer would take to request a cluster, but without using the UI or
having to wait for a full cluster install.

This script takes a multi-threaded approach in order to run the various
scenarios in parallel.