Updates the validation controller OU logic
Changes the validation controller to not care if the account has been
reused before, but only care if the legal entity id is currently set.
This allows us to take into consideration accounts that have had the
legal entity reset and have not been claimed again, because the `reused`
status is still set to true on those accounts.