OSD-18390: Ensure Region Initialization Fails on Missing AMI Currently, when a region is being initialized, if we fail to lookup an AMI to use, or the instance type, the code would continue to execute and later fail on passing an empty string to AWS.
The function performing the initialization does return errors as well, but they are never read by the controller calling. Instead it spawns a goroutine for each region init, and uses channels to handle errors and info messages.

In the scenario where the AMI lookup fails, the error is sent back on the errors channel, and then that channel is closed, following by the goroutine continuing to run in the background until it eventually panics trying to send to the now closed error channel.

This causes the operator process to exit prematurely, leading to odd scenarios around cleaning up initialization VPCs that get leaked.